BUGFIX: Call gtk.gdk.threads_enter() before gtk.main().
After calling gtk.gdk.threads_init(), a lock is created which the caller is
expected to acquire before calling gtk.main() which subsequently releases this
lock. In previous versions of GTK/GDK, performing a release on an unlocked lock
didn't cause problems and so the missing threads_enter() call did not crash
things. At some point this changed causing Perentie to crash on start. This
patch adds a call to threads_enter() before calling gtk.main() preventing this
bug.

NB: In later versions of GTK a workaround in GTK/GDK was added because it
turned out that a huge number of applications contained this same bug.

For more details see:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=758619#167

And other messages in that thread...